refactor: Consolidate duplicate code into shared utilities
- Create frontend/src/utils/format.ts with centralized formatting functions
  (formatFileSize, formatRelativeTime, formatDownloadCount, formatDate, formatDuration)
- Create frontend/src/constants/packages.ts for packageTypeLabels mapping
- Update 8 frontend components to use shared utilities instead of duplicate implementations
- Add repo_to_response() helper in backend repositories handler to reduce repetition

Removes ~250 lines of duplicate code across the codebase.
