+++ frontend/src/components/builds/BuildList/BuildList.tsx
@@ -12,6 +12,7 @@ import {
 import type { ColumnsType, TablePaginationConfig } from 'antd/es/table';
 import type { Build, BuildStatus } from '../../../types';
 import { colors } from '../../../styles/tokens';
+import { formatDate, formatDuration } from '../../../utils';
 
 const { Text } = Typography;
 
@@ -26,33 +27,6 @@ export interface BuildListProps {
     onChange: (page: number, pageSize: number) => void;
   };
 }
-
-const formatDate = (dateString: string | undefined): string => {
-  if (!dateString) return '-';
-  const date = new Date(dateString);
-  return date.toLocaleString(undefined, {
-    year: 'numeric',
-    month: 'short',
-    day: 'numeric',
-    hour: '2-digit',
-    minute: '2-digit',
-  });
-};
-
-const formatDuration = (durationMs: number | undefined): string => {
-  if (!durationMs) return '-';
-  const seconds = Math.floor(durationMs / 1000);
-  const minutes = Math.floor(seconds / 60);
-  const hours = Math.floor(minutes / 60);
-
-  if (hours > 0) {
-    return `${hours}h ${minutes % 60}m`;
-  }
-  if (minutes > 0) {
-    return `${minutes}m ${seconds % 60}s`;
-  }
-  return `${seconds}s`;
-};
 
 const getStatusIcon = (status: BuildStatus): React.ReactNode => {
   const iconStyle = { marginRight: 4 };
