+++ frontend/src/components/dashboard/widgets/RecentActivityWidget.tsx
@@ -12,6 +12,7 @@ import { useNavigate } from 'react-router-dom';
 import { DashboardWidget } from '../DashboardWidget';
 import { artifactsApi, repositoriesApi } from '../../../api';
 import { colors, spacing } from '../../../styles/tokens';
+import { formatRelativeTime } from '../../../utils';
 import type { Artifact } from '../../../types';
 
 const { Text, Link } = Typography;
@@ -43,22 +44,6 @@ const activityColors: Record<ActivityType, string> = {
   upload: colors.success,
   download: colors.info,
   delete: colors.error,
-};
-
-const formatRelativeTime = (dateString: string): string => {
-  const date = new Date(dateString);
-  const now = new Date();
-  const diffMs = now.getTime() - date.getTime();
-  const diffSecs = Math.floor(diffMs / 1000);
-  const diffMins = Math.floor(diffSecs / 60);
-  const diffHours = Math.floor(diffMins / 60);
-  const diffDays = Math.floor(diffHours / 24);
-
-  if (diffSecs < 60) return 'Just now';
-  if (diffMins < 60) return `${diffMins} minute${diffMins > 1 ? 's' : ''} ago`;
-  if (diffHours < 24) return `${diffHours} hour${diffHours > 1 ? 's' : ''} ago`;
-  if (diffDays < 7) return `${diffDays} day${diffDays > 1 ? 's' : ''} ago`;
-  return date.toLocaleDateString();
 };
 
 export const RecentActivityWidget: React.FC = () => {
